refactor: Application name, Country.kt 변경 사항  - 애플리케이션 이름을 doume에서 doumi로 수정함  - Country.kt의 @Getter 어노테이션을 삭제함.
