+++ src/main/kotlin/hs/kr/gbsw/doumi/user/model/Country.kt
@@ -1,4 +1,4 @@
-package hs.kr.gbsw.doume.user.model
+package hs.kr.gbsw.doumi.user.model
 
 import jakarta.persistence.Entity
 import jakarta.persistence.GeneratedValue
@@ -6,7 +6,6 @@ import jakarta.persistence.GenerationType
 import jakarta.persistence.Id
 import lombok.Getter
 
-@Getter
 @Entity
 class Country(
     @Id
