+++ src/main/kotlin/hs/kr/gbsw/doumi/user/repository/UserRepository.kt
@@ -1,6 +1,6 @@
-package hs.kr.gbsw.doume.user.repository
+package hs.kr.gbsw.doumi.user.repository
 
-import hs.kr.gbsw.doume.user.model.Users
+import hs.kr.gbsw.doumi.user.model.Users
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface UserRepository: JpaRepository<Users, Long> {
