--- src/test/kotlin/hs/kr/gbsw/doume/DoumeApplicationTests.kt
@@ -1,19 +1,0 @@
-package hs.kr.gbsw.doume
-
-import hs.kr.gbsw.doume.user.model.Country
-import hs.kr.gbsw.doume.user.model.Users
-import hs.kr.gbsw.doume.user.repository.CountryRepository
-import hs.kr.gbsw.doume.user.repository.UserRepository
-import org.junit.jupiter.api.Test
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class DoumeApplicationTests {
-
-	@Test
-	fun test() {
-
-	}
-
-}
