+++ src/test/kotlin/hs/kr/gbsw/doumi/DoumeApplicationTests.kt
@@ -1,0 +1,14 @@
+package hs.kr.gbsw.doumi
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class DoumeApplicationTests {
+
+	@Test
+	fun test() {
+
+	}
+
+}
